src/main/java/org/harpreet/internetmeme/MemeRecorder.java
/*
 * Copyright (c) 2009 Cliffano Subagio
 * Copyright (c) 2013 Harpreet Singh
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package org.harpreet.internetmeme;

import hudson.FilePath;
import hudson.Launcher;
import hudson.model.*;
import hudson.tasks.BuildStepMonitor;
import hudson.tasks.Recorder;

import jenkins.tasks.SimpleBuildStep;
import org.kohsuke.stapler.DataBoundConstructor;

import javax.annotation.Nonnull;

/**
 * This class associates a MemeAction to a job or a build.
 * @author clifanno
 * @author Harpreet Singh
 */
public class MemeRecorder extends Recorder implements SimpleBuildStep {

    /**
     * Constructs a {@link MemeRecorder}
     */
    @DataBoundConstructor
    public MemeRecorder() {
    }

    /**
     * Gets the MemeAction as the project action. This is applicable for
     * each job and only when there's at least one build in the job.
     * @param project
     *            the project
     * @return the project action
     */
    @Override
    public final Action getProjectAction(final AbstractProject project) {
        Action action = null;
        if (project.getLastBuild() != null) {
            Style style = Style.get(project.getLastBuild().getResult());
            action = new MemeAction(style);
        }
        return action;
    }

    /**
     * Adds MemeAction to the build actions. This is applicable for each
     * build.
     * @param build
     *            the build
     * @param launcher
     *            the launcher
     * @param listener
     *            the listener
     * @return true
     */
    @Override
    public final boolean perform(final AbstractBuild build,
            final Launcher launcher, final BuildListener listener) {
        perform(build);
        return true;
    }

    /**
     * Adds MemeAction to the run actions. This is applicable for each
     * run.
     * @param run
     *            the run
     * @param workspace
     *            the workspace
     * @param launcher
     *            the launcher
     * @param listener
     *            the listener
     */
    @Override
    public final void perform(
            @Nonnull final Run run, @Nonnull final FilePath workspace,
            @Nonnull final Launcher launcher, @Nonnull final TaskListener listener) {
        perform(run);
    }

    /**
     * Gets the required monitor service.
     * @return the BuildStepMonitor
     */
    public final BuildStepMonitor getRequiredMonitorService() {
        return BuildStepMonitor.NONE;
    }

    /**
     * Adds MemeAction to the run actions. This is applicable for each
     * run.
     * @param run
     *            the run
     */
    public void perform(final Run<?, ?> run) {
        Style style = Style.get(run.getResult());
        run.addAction(new MemeAction(style));
    }
}
